rename: 더미데이터 폴더 이름 변경
