+++ src/components/Sections/Latest.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import ContentsCard from '../Cards/ContentsCard'
-import { latestData } from '@/dummys/section-data'
+import { latestData } from '@/dummies/section-data'
 
 export default function Latest() {
   return (
